refactor(provider): extract Benchmark into its own ProviderBenchmark target
Benchmark/ (~5.2k LOC, incl. fatalError protocol-conformance stub caches)
was compiled into the ProviderCore library and thus linked into every
provider/enclave binary. Move it into a dedicated `ProviderBenchmark`
SwiftPM target that depends on ProviderCore, so only the benchmark-bearing
executables link it (darkbloom `benchmark`, kv-quant-gate, kv-attn-selftest,
kv-engine-demo) and the shipped enclave/publish/core surface no longer
carries benchmark code.

- new target `ProviderBenchmark` (Sources/ProviderBenchmark), 18 files moved
- the two engine-facing types the live scheduler needs stay in ProviderCore:
  KVQuantCandidateMode (split out of KVQuantTypes.swift) and the
  self-contained KVQuantPolicy.swift, now under ProviderCore/Inference/KVQuant
- widen KVEstimation / parseModelArchitecture / resolvedKVBytesPerToken /
  ModelArchitecture to public (already used by the live scheduler) so the
  benchmark target can reach them across the module boundary
- wire ProviderBenchmark into the four executables + the four benchmark
  test files; add `import ProviderCore` to the moved files

Behavior-preserving: swift build green; swift test 1049 pass / 0 fail.
