+++ provider-swift/Sources/ProviderBenchmark/KVQuant/KVQuantCache.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ProviderCore
 import MLX
 import MLXLMCommon
 import MLXNN
