+++ provider-swift/Sources/ProviderBenchmark/KVQuant/KVQuantGateRunner.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ProviderCore
 import ProviderCoreFoundation
 
 public struct KVQuantGateRunner {
