+++ provider-swift/Sources/ProviderBenchmark/KVQuant/KVQuantQualityRunners.swift
@@ -1,4 +1,5 @@
 import Foundation
+import ProviderCore
 
 public struct KVQuantQualityRunner {
     public init() {}
